Cambios y permisos

+++ apps/core/view/personal/views.py
@@ -3,6 +3,7 @@ from django.contrib.auth.mixins import LoginRequiredMixin
 from django.utils.decorators import method_decorator
 from django.views.decorators.csrf import csrf_exempt
 from django.http import JsonResponse
+from apps.core.mixins import Perms_Check
 
 # IMPORTACIONES DE LOS MODELOS Y LOS FORMULARIOS
 from apps.core.models import Personal
@@ -12,7 +13,8 @@ from apps.core.forms import PersonalForm
 # Create your views here.
 
 # PERSONAL 
-class PersonalViews(LoginRequiredMixin, TemplateView):
+class PersonalViews(Perms_Check, LoginRequiredMixin, TemplateView):
+	permission_required = 'core.view_personal'
 	template_name =  'personal/ListadoPersonal.html'
 
 	@method_decorator(csrf_exempt)
@@ -25,47 +27,36 @@ class PersonalViews(LoginRequiredMixin, TemplateView):
 			action = request.POST['action']
 			
 			if action == 'listado_personal':
-				data = []
-				for i in Personal.objects.all():
-					data.append(i.toJSON())
+				perms = ('core.view_personal',)
+				if request.user.has_perms(perms):
+					data = []
+					for i in Personal.objects.all():
+						data.append(i.toJSON())
 
 			elif action == 'agregar_personal':
-				if Personal.objects.filter(cedula = request.POST.get('cedula')):
-					data['error'] = 'Ya existe personal registrado con este numero de cédula'   
-				else:
-					per = Personal()
-					print(request.POST)
-					per.cedula = request.POST.get('cedula')
-					per.nombre = request.POST.get('nombre')
-					per.apellido = request.POST.get('apellido')
-					per.direccion = request.POST.get('direccion')
-					per.movil = request.POST.get('movil')
-					per.correo = request.POST.get('correo')
-					per.ocupacion = request.POST.get('ocupacion')
-					per.sexo = request.POST.get('sexo')
-					per.status = request.POST.get('status')
-					per.save()
-
-			elif action == 'editar_personal':
-			
-				per = Personal.objects.get(pk = request.POST['id'])
-				if per.cedula == request.POST.get('cedula'):
-					per.nombre = request.POST.get('nombre')
-					per.apellido =request.POST.get('apellido')
-					per.direccion = request.POST.get('direccion')
-					per.movil = request.POST.get('movil')
-					per.correo = request.POST.get('correo')
-					if request.POST.get('ocupacion'):
-						per.ocupacion = request.POST.get('ocupacion')
-					else:
-						per.ocupacion = per.ocupacion
-					per.status = request.POST.get('status')
-					per.save()
-				else:
+				perms = ('core.add_personal',)
+				if request.user.has_perms(perms):
 					if Personal.objects.filter(cedula = request.POST.get('cedula')):
 						data['error'] = 'Ya existe personal registrado con este numero de cédula'   
 					else:
+						per = Personal()
+						print(request.POST)
 						per.cedula = request.POST.get('cedula')
+						per.nombre = request.POST.get('nombre')
+						per.apellido = request.POST.get('apellido')
+						per.direccion = request.POST.get('direccion')
+						per.movil = request.POST.get('movil')
+						per.correo = request.POST.get('correo')
+						per.ocupacion = request.POST.get('ocupacion')
+						per.sexo = request.POST.get('sexo')
+						per.status = request.POST.get('status')
+						per.save()
+
+			elif action == 'editar_personal':
+				perms = ('core.change_personal',)
+				if request.user.has_perms(perms):
+					per = Personal.objects.get(pk = request.POST['id'])
+					if per.cedula == request.POST.get('cedula'):
 						per.nombre = request.POST.get('nombre')
 						per.apellido =request.POST.get('apellido')
 						per.direccion = request.POST.get('direccion')
@@ -77,20 +68,42 @@ class PersonalViews(LoginRequiredMixin, TemplateView):
 							per.ocupacion = per.ocupacion
 						per.status = request.POST.get('status')
 						per.save()
+					else:
+						if Personal.objects.filter(cedula = request.POST.get('cedula')):
+							data['error'] = 'Ya existe personal registrado con este numero de cédula'   
+						else:
+							per.cedula = request.POST.get('cedula')
+							per.nombre = request.POST.get('nombre')
+							per.apellido =request.POST.get('apellido')
+							per.direccion = request.POST.get('direccion')
+							per.movil = request.POST.get('movil')
+							per.correo = request.POST.get('correo')
+							if request.POST.get('ocupacion'):
+								per.ocupacion = request.POST.get('ocupacion')
+							else:
+								per.ocupacion = per.ocupacion
+							per.status = request.POST.get('status')
+							per.save()
 			
 			elif action == 'delete_personal':
-				per = Personal.objects.get(id = request.POST.get('id'))
-				per.delete()
+				perms = ('core.delete_personal',)
+				if request.user.has_perms(perms):
+					per = Personal.objects.get(id = request.POST.get('id'))
+					per.delete()
 
 			elif action == 'activar_personal':
-				per = Personal.objects.get(id = request.POST.get('id'))
-				per.status = 'Activo'
-				per.save()
+				perms = ('core.change_personal',)
+				if request.user.has_perms(perms):
+					per = Personal.objects.get(id = request.POST.get('id'))
+					per.status = 'Activo'
+					per.save()
 
 			elif action == 'desactivar_personal':
-				per = Personal.objects.get(id = request.POST.get('id'))
-				per.status = 'Inactivo'
-				per.save()
+				perms = ('core.change_personal',)
+				if request.user.has_perms(perms):
+					per = Personal.objects.get(id = request.POST.get('id'))
+					per.status = 'Inactivo'
+					per.save()
 				
 			else:
 				data['error'] = 'Ha ocurrido un error'
